+++ src/pages/docs/getting-started/overview.mdx
@@ -1,57 +1,34 @@
-import { Lead } from "./docs"
+import Layout from "../../../components/docs-layout"
+export default Layout
 
 # Overview
 
-<Lead>
-  Mirage is a library that lets frontend developers mock out backend APIs.
-</Lead>
+Mirage lets you simulate API responses by writing **route handlers**.
 
-<Lead>
-  Unlike other mocking libraries, Mirage makes it easy to recreate dynamic
-  scenarios that are typically only possible with real production servers.
-</Lead>
-
----
-
-## Installation
-
-Add Mirage.js to your project by running
-
-```shell
-# Using Yarn
-yarn add --dev @miragejs/server
-
-# Using npm
-npm install --save-dev @miragejs/server
-```
-
-## Usage
-
-Mirage lets you simulate API responses by writing **route handlers**. The simplest example of a route handler is a function that returns an object:
+The simplest example of a route handler is a function that returns an object:
 
 ```js
-import { Server } from "@miragejs/server"
+// mirage/config.js
+export default function() {
+  this.namespace = "api"
 
-new Server({
-  routes(server) {
-    // Define your API endpoints in this function
-
-    server.get("/movies", () => {
-      return [
-        { id: 1, type: "movies", name: "Interstellar" },
-        { id: 2, type: "movies", name: "Inception" },
-        { id: 3, type: "movies", name: "Dunkirk" },
-      ]
-    })
-  },
-})
+  this.get("/movies", () => {
+    return {
+      data: [
+        { id: 1, type: "movies", attributes: { name: "Interstellar" } },
+        { id: 2, type: "movies", attributes: { name: "Inception" } },
+        { id: 3, type: "movies", attributes: { name: "Dunkirk" } },
+      ],
+    }
+  })
+}
 ```
 
-Now whenever your JavaScript app makes a GET request to `/api/movies`, Mirage will respond with this data.
+Now whenever your Ember app makes a GET request to `/api/movies`, Mirage will respond with this data.
 
 ## Dynamic data
 
-This works, and is a common way to simulate HTTP responses, but hard-coded responses like this have a few problems:
+This works, and is a common way to simulate HTTP responses - but hard-coded responses like this have a few problems:
 
 - _They're inflexible_. What if you want to change the data for this route in your tests?
 - _They contain formatting logic_. Logic that's concerned with the shape of your JSON payload (e.g. the `data` and `attributes` keys) is now duplicated across all your route handlers.
